requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-text-splitters
langchain-community
langchain-ollama
faiss-cpu
ollama
python-dotenv
ruff
pytest
